+++ cli/lib/kontena/cli/apps/service_generator.rb
@@ -22,15 +22,36 @@ module Kontena::Cli::Apps
 
     private
 
+    # @param [Hash] options
+    def merge_env_vars(options)
+      return options['environment'] unless options['env_file']
+
+      options['env_file'] = [options['env_file']] if options['env_file'].is_a?(String)
+      options['environment'] = [] unless options['environment']
+
+      options['env_file'].each do |env_file|
+        options['environment'].concat(read_env_file(env_file))
+      end
+      options['environment'].uniq {|s| s.split('=').first}
+    end
+
+
+    # @param [String] path
+    def read_env_file(path)
+      File.readlines(path).delete_if { |line| line.start_with?('#') || line.empty? }
+    end
+
     ##
     # @param [Hash] options
     # @return [Hash]
     def parse_data(options)
+
       data = {}
       data[:image] = parse_image(options['image'])
-      data[:env] = options['environment']
+      data[:env] = merge_env_vars(options)
       data[:container_count] = options['instances']
       data[:links] = parse_links(options['links'] || [])
+      data[:external_links] = parse_links(options['external_links'] || [])
       data[:ports] = parse_ports(options['ports'] || [])
       data[:memory] = parse_memory(options['mem_limit'].to_s) if options['mem_limit']
       data[:memory_swap] = parse_memory(options['memswap_limit'].to_s) if options['memswap_limit']
