+++ cli/lib/kontena/cli/apps/stop_command.rb
@@ -17,7 +17,7 @@ module Kontena::Cli::Apps
       require_config_file(filename)
 
       @service_prefix = project_name || current_dir
-      @services = load_services(filename, service_list, service_prefix)
+      @services = services_from_yaml(filename, service_list, service_prefix)
       if services.size > 0
         stop_services(services)
       elsif !service_list.empty?
