+++ cli/spec/kontena/cli/app/common_spec.rb
@@ -30,122 +30,32 @@ describe Kontena::Cli::Apps::Common do
     }
   end
 
-  describe '#parse_services' do
-
-    it 'returns services from given YAML file' do
+  describe '#load_from_yaml' do
+    before(:each) do
       allow(File).to receive(:read).with("#{Dir.getwd}/kontena.yml").and_return(kontena_yml)
       allow(File).to receive(:read).with("#{Dir.getwd}/docker-compose.yml").and_return(docker_compose_yml)
-      services = subject.parse_services('kontena.yml')
+    end
+    it 'populate env variables' do
+      services = subject.services_from_yaml('kontena.yml',[],'load-test')
+      expect(ENV['grid']).to eq('test-grid')
+      expect(ENV['project']).to eq('load-test')
+    end
+
+    it 'returns services from given YAML file' do
+      services = subject.services_from_yaml('kontena.yml',[],'')
       expect(services['wordpress']).not_to be_nil
     end
 
-    it 'raises error if extended service is not found from base file' do
-      allow(File).to receive(:read).with("#{Dir.getwd}/kontena.yml").and_return(kontena_yml)
-      allow(File).to receive(:read).with("#{Dir.getwd}/docker-compose.yml").and_return(mysql_yml)
-
-      expect {
-        subject.parse_services('kontena.yml')
-      }.to raise_error(SystemExit)
-    end
-  end
-
-  describe '#normalize_env_vars' do
-    it 'converts env hash to array' do
-      opts = {
-          'environment' => {
-              'FOO' => 'bar',
-              'BAR' => 'baz'
-          }
-      }
-      subject.normalize_env_vars(opts)
-      env = opts['environment']
-      expect(env).to include('FOO=bar')
-      expect(env).to include('BAR=baz')
+    it 'aborts on validation failure' do
+      allow_any_instance_of(Kontena::Cli::Apps::YAML::Validator).to receive(:validate)
+        .and_return([{ 'wordress' => [] } ])
+      expect { subject.services_from_yaml('kontena.yml',[],'') }.to raise_error(SystemExit)
     end
 
-    it 'does nothing to env array' do
-      opts = {
-          'environment' => [
-              'FOO=bar', 'BAR=baz'
-          ]
-      }
-      subject.normalize_env_vars(opts)
-      env = opts['environment']
-      expect(env).to include('FOO=bar')
-      expect(env).to include('BAR=baz')
-    end
-  end
-
-  describe '#extend_env_vars' do
-    it 'inherites env vars from upper level' do
-      from = ['FOO=bar']
-      to = nil
-      env_vars = subject.extend_env_vars(from, to)
-      expect(env_vars).to eq(['FOO=bar'])
-    end
-
-    it 'overrides values' do
-      from = ['FOO=bar']
-      to = ['FOO=baz']
-      env_vars = subject.extend_env_vars(from, to)
-      expect(env_vars).to eq(['FOO=baz'])
-    end
-
-    it 'combines variables' do
-      from = ['FOO=bar']
-      to = ['BAR=baz']
-      env_vars = subject.extend_env_vars(from, to)
-      expect(env_vars).to eq(['BAR=baz', 'FOO=bar'])
-    end
-  end
-
-  describe '#extend_secrets' do
-    it 'inherites secrets from upper level' do
-      secret = {
-        'secret' => 'CUSTOMER_DB_PASSWORD',
-        'name' => 'MYSQL_PASSWORD',
-        'type' => 'env'
-      }
-      from = [secret]
-      to = nil
-      secrets = subject.extend_secrets(from, to)
-      expect(secrets).to eq([secret])
-    end
-
-    it 'overrides secrets' do
-      from_secret = {
-        'secret' => 'CUSTOMER_DB_PASSWORD',
-        'name' => 'MYSQL_PASSWORD',
-        'type' => 'env'
-      }
-
-      to_secret = {
-        'secret' => 'CUSTOMER_DB_PASSWORD',
-        'name' => 'MYSQL_ROOT_PASSWORD',
-        'type' => 'env'
-      }
-      from = [from_secret]
-      to = [to_secret]
-      secrets = subject.extend_secrets(from, to)
-      expect(secrets).to eq([to_secret])
-    end
-
-    it 'combines secrets' do
-      from_secret = {
-        'secret' => 'CUSTOMER_DB_PASSWORD',
-        'name' => 'MYSQL_PASSWORD',
-        'type' => 'env'
-      }
-
-      to_secret = {
-        'secret' => 'CUSTOMER_API_TOKEN',
-        'name' => 'API_TOKEN',
-        'type' => 'env'
-      }
-      from = [from_secret]
-      to = [to_secret]
-      secrets = subject.extend_secrets(from, to)
-      expect(secrets).to eq([to_secret, from_secret])
+    it 'returns given service from given YAML file' do
+      services = subject.services_from_yaml('kontena.yml',['wordpress'],'')
+      expect(services['wordpress']).not_to be_nil
+      expect(services.size).to eq(1)
     end
   end
 end
